test: additional tests for routing a mock event
