+++ htsget-elsa-lambda/src/config.rs
@@ -9,28 +9,40 @@ use tracing::instrument;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Config {
-    #[serde(flatten)]
+    #[serde(flatten, default)]
     htsget_config: HtsGetConfig,
     #[serde(with = "http_serde::authority")]
-    elsa_endpoint: Authority,
+    elsa_endpoint_authority: Authority,
     cache_location: String,
 }
 
 impl Config {
-    #[instrument(level = "debug", ret)]
-    pub fn from_path(path: &Path) -> io::Result<Self> {
-        from_path(path)
+    pub fn new(
+        htsget_config: HtsGetConfig,
+        elsa_endpoint_authority: Authority,
+        cache_location: String,
+    ) -> Self {
+        Self {
+            htsget_config,
+            elsa_endpoint_authority,
+            cache_location,
+        }
     }
 
     pub fn htsget_config(&self) -> &HtsGetConfig {
         &self.htsget_config
     }
 
-    pub fn elsa_endpoint(&self) -> &Authority {
-        &self.elsa_endpoint
+    pub fn elsa_endpoint_authority(&self) -> &Authority {
+        &self.elsa_endpoint_authority
     }
 
     pub fn cache_location(&self) -> &str {
         &self.cache_location
+    }
+
+    #[instrument(level = "debug", ret)]
+    pub fn from_path(path: &Path) -> io::Result<Self> {
+        from_path(path)
     }
 }
